requirements.txt
pandas
openai
python-dotenv
emoji
fastapi
uvicorn
pytest